feat: include payment summary
