+++ packages/app/src/containers/TicketSellingPage/molecules/TicketTableBody.js
@@ -6,7 +6,7 @@
 /* System imports */
 import React, { PropTypes } from 'react';
 import { formatDescription } from 'utils/ticket';
-import { calculateSubtotal } from 'selectors/tickets';
+import { getSubtotal, formatDecimalPlaces } from 'utils/ticket';
 import TicketTableBodyContainer from '../atoms/TicketTableBodyContainer';
 import HeightAdapterContainer from '../atoms/HeightAdapterContainer';
 import TicketTableRowContainer from '../atoms/TicketTableRowContainer';
@@ -65,7 +65,7 @@ export class TicketTableBody extends React.Component {
                   }
                 }}
               />
-              <TicketTableField placeholder={calculateSubtotal(operation).toFixed(2)} readonly />
+              <TicketTableField placeholder={formatDecimalPlaces(getSubtotal(operation))} readonly />
               <TicketTableButton primary icon="remove-close" onClick={() => this.props.removeStockFromTicket(i)} />
             </TicketTableRowContainer>
           ))}
